funcion de crear productos

+++ src/main/java/com/fontebo/inventory/Models/Category.java
@@ -16,8 +16,8 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    
+    private Long id;
     private String name;
